requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-docx
pdfplumber
pytesseract
faiss-cpu
openai
python-dotenv
pandas